+++ app/src/main/java/omikhieiev/app/domain/boundaries/app/TransactionsUseCase.kt
@@ -8,6 +8,6 @@ interface TransactionsUseCase {
     suspend fun getAllTransactions(): Result<List<Transaction>>
 
     suspend fun calculateAverage(transactions: List<Transaction>): Double
-    suspend fun calculateSum(transactions: List<Transaction>): Int
+    suspend fun calculateSum(transactions: List<Transaction>): Double
 
 }
