+++ app/src/main/java/omikhieiev/app/ui/main/GeneralInfo.kt
@@ -1,4 +1,4 @@
 package omikhieiev.app.ui.main
 
-data class GeneralInfo (val sum: Int, val average: Double) {
+data class GeneralInfo (val sum: Double, val average: Double) {
 }
